requirements.txt
whisper
pytubefix
opencv-python
ffmpeg-python